Implement choosing of the parent folder to name a job directory.
Currently, the job directory either gets the name of the current parent folder (which is the dropzone value if the dropped element is a folder and the parent if the dropped element is a file) or a timestamp if no dropzone is part of a plugin.

Next step will be to check for an existing 0.info file and parse it for alternative naming.