Create a chris.run file which executes the command and the chmod -R 755 when the command is done.
This was required since mosbatch only accepts one program to run. An advantage is that a job can be easily run again using this (executable) file.

I do observe some slow performance in job scheduling which might be a temporary thing?

Also, we still need to fix the chmod 777 whenever a feed is shared.