Reduce the SSH logins for one job submission to 3.
This makes it much faster.

A chris.param file is not needed anymore since we have chris.run.

Also, I think it might be possible to combine it all to maximal 2 ssh logins when we create the job directoy not before launching the job.

This way it might be possible to detect whether a job was "queued" or is actually "running".